make AwsWrapperDbContextOptionsExtensionInfo private sealed class
